request for rectification(KaTalk text)
