+++ TalkSsogi_server/src/main/java/com/talkssogi/TalkSsogi_server/domain/AnalysisResult.java
@@ -14,7 +14,6 @@ public class AnalysisResult {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @Id
     private String chatroomName;
 
     @OneToOne // 일대일
