+++ TalkSsogi_server/src/main/java/com/talkssogi/TalkSsogi_server/service/AnalysisResultService.java
@@ -24,8 +24,8 @@ public class AnalysisResultService {
         return analysisResultRepository.findByWordCloudImageUrl(wordCloudImageUrl);
     }
 
-    public String findWordCloudImageUrlByChatRoomIdAndUserId(Integer chatRoomId, Integer userId) {
-        AnalysisResult analysisResult = analysisResultRepository.findByChattingRoom_CrNumAndChattingRoom_User_Id(chatRoomId, userId);
+    public String findWordCloudImageUrlByChatRoomIdAndUserId(Integer chatRoomId, String userId) {
+        AnalysisResult analysisResult = analysisResultRepository.findByChattingRoom_CrNumAndChattingRoom_User_UserId(chatRoomId, userId);
         if (analysisResult != null) {
             return analysisResult.getWordCloudImageUrl();
         }
